Day02 part A redifine IdsValidator and tested

+++ src/main/java/software/aoc/day02/a/model/Id.java
@@ -1,4 +1,4 @@
 package software.aoc.day02.a.model;
 
-public record Id(int id) {
+public record Id(long id) {
 }
